IGNITE-20443 QueryHasNoMoreResultsException moved to public API.
